Actions fixed - Async/Await funtions corrected - New functions in api folder, auth file
